Factor self-connected Tahoe-LAFS storage node code into a fixture
